test(button): migrate form-reference
